main.js подключен а index.html как модуль, скрипты разбиты на модули, добавила пару пустых файлов-модулей по тз
