+++ js/map.js
@@ -1,0 +1,1 @@
+//модуль карты
